Separates model exceptions.
