+++ src/model.cpp
@@ -1,10 +1,10 @@
 #include "model.hpp"
 #include "exceptions/load.hpp"
+#include "exceptions/model.hpp"
 #include "image_loader.hpp"
 #include "linear.hpp"
 #include "metrics.hpp"
 #include "utils/cli.hpp"
-#include "utils/exceptions.hpp"
 #include "utils/indicator.hpp"
 #include "utils/math.hpp"
 #include "utils/string.hpp"
