+++ src/utils/exceptions.hpp
@@ -5,100 +5,6 @@
 #include <vector>
 
 namespace exceptions {
-#pragma region Model
-namespace model {
-class InvalidTotalEpochException : public std::exception {
-  int totalEpochs;
-
-  virtual const char *what() const throw() {
-    std::string s = "Total epochs must be 0 or greater. Got: " +
-                    std::to_string(totalEpochs) + ".";
-    char *result = new char[s.length() + 1];
-    std::strcpy(result, s.c_str());
-    return result;
-  }
-
-public:
-  InvalidTotalEpochException(int totalEpochs) : totalEpochs(totalEpochs){};
-};
-
-class InvalidMetricException : public std::exception {
-  std::string metric;
-
-  virtual const char *what() const throw() {
-    std::string s = metric + " is not a valid metric.";
-    char *result = new char[s.length() + 1];
-    std::strcpy(result, s.c_str());
-    return result;
-  }
-
-public:
-  InvalidMetricException(std::string metric) : metric(metric){};
-};
-
-class EmptyLayersVectorException : public std::exception {
-  virtual const char *what() const throw() {
-    return "Layers vector cannot be empty.";
-  }
-};
-
-class MissingClassesException : public std::exception {
-  virtual const char *what() const throw() {
-    return "Model is missing the classes.";
-  }
-};
-
-class ClassHistoryMismatchException : public std::exception {
-  int classSize, historySize;
-  std::string metric;
-
-  virtual const char *what() const throw() {
-    std::string s =
-        "The number of classes (" + std::to_string(this->classSize) +
-        ") does not match the number of elements for \"" + this->metric +
-        "\" (" + std::to_string(this->historySize) + ").";
-    char *result = new char[s.length() + 1];
-    std::strcpy(result, s.c_str());
-    return result;
-  }
-
-public:
-  ClassHistoryMismatchException(int classSize, int historySize,
-                                const std::string &metric)
-      : classSize(classSize), historySize(historySize), metric(metric){};
-};
-
-class InvalidExtensionException : public std::exception {
-  std::string extension;
-
-  virtual const char *what() const throw() {
-    std::string s = "File format \"" + this->extension + "\" is not supported.";
-    char *result = new char[s.length() + 1];
-    std::strcpy(result, s.c_str());
-    return result;
-  }
-
-public:
-  InvalidExtensionException(const std::string &extension)
-      : extension(extension){};
-};
-
-class InvalidPlottingMetricException : public std::exception {
-  std::string metric;
-
-  virtual const char *what() const throw() {
-    std::string s = "Plotting \"" + this->metric + "\" is not supported.";
-    char *result = new char[s.length() + 1];
-    std::strcpy(result, s.c_str());
-    return result;
-  }
-
-public:
-  InvalidPlottingMetricException(const std::string &metric) : metric(metric){};
-};
-} // namespace model
-#pragma endregion Model
-
 #pragma region Utils
 namespace utils {
 #pragma region One hot encode
